+++ CHANGELOG.md
@@ -87,6 +87,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Prevent users from modifying a group with permissions they don't hold - #5380 by @IKarbowiak
 - Add user_permissions on user type - #5390 by @IKarbowiak
 - Update user to group assign restriction - #5395 by @IKarbowiak
+- Add editable groups on user type - #5408 by @IKarbowiak
 
 ## 2.9.0